feature: criando estrutura inicial pra comunicação com o gateway de integração

+++ src/main/java/com/client/ws/rasmooplus/dto/wsraspay/OrderDto.java
@@ -1,4 +1,6 @@
 package com.client.ws.rasmooplus.dto.wsraspay;
 
-public record OrderDto(String id, String customerId, Long discount, String productAcronym) {
+import java.math.BigDecimal;
+
+public record OrderDto(String id, String customerId, BigDecimal discount, String productAcronym) {
 }
